Bump the version of flickr-url-parser
This needs two new test cassettes because there was a slight behaviour
change: it now returns user URLs with a trailing slash, for consistency
with how Flickr.com does it.

I thought there'd be a way to simplify the photos-api lookups also, but
that doesn't seem to be true -- we need at least one API call to look up
the user info, and lookupUserByUrl works well enough for that.