Create a new helper `parse_timestamp`
This is functionally the same as the old helper, but with a slightly
more generic name to reflect the fact it's being used to parse timestamps
from lots of parts of the Flickr API, not just "date posted" on photos.

I'm also breaking up the `utils.py` file, which is a poorly-named file --
there's enough date-related code to belong in its own file.