Skip a test that fails if FLICKR_API_KEY is set
If you run this test while you have FLICKR_API_KEY set, previously it
would make the real API call and the test would fail because it expects
to get a configuration error. By skipping the test when the env var is
set, we avoid creating spurious VCR cassettes and getting unexpected
failures.